Managed to figure out use model container in preview
